% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64_to_df.R
\name{base64_to_df}
\alias{base64_to_df}
\title{base64_to_df}
\usage{
base64_to_df(x)
}
\arguments{
\item{x}{The base64-encoded CSV string}
}
\value{
A `data.frame` object containing the data from `x`.
}
\description{
Convert a base64 representation of a CSV table into a
`data.frame` object.
}
