% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesurvey_api_call.R
\name{limesurvey_api_call}
\alias{limesurvey_api_call}
\title{limesurvey_api_call}
\usage{
limesurvey_api_call(method, params = list(), ...)
}
\arguments{
\item{method}{Name of the API method to call.
A complete list of methods can be found here:
https://api.limesurvey.org/classes/remotecontrol_handle.html}

\item{params}{Parameters to pass to the API}

\item{...}{Additional parameters passed from above}
}
\value{
A list containing the de-serialized response.
}
\description{
Perform a call to the `LimeSurvey` RPC API.
}
