% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wipe_survey_data.R
\name{wipe_survey_data}
\alias{wipe_survey_data}
\title{wipe_survey_data}
\usage{
wipe_survey_data(survey_id)
}
\arguments{
\item{survey_id}{ID of the survey from which the collected data shall be
deleted.
6-digit integer.}
}
\value{
Nothing. Function is called for side effects on SQL table.
}
\description{
Delete all data collected by this survey.
}
\examples{
# This example assumes a locally hosted `LimeSurvey` instance using a locally
# hosted MySQL server.
# On this `LimeSurvey` instance, there is a survey with the ID 123456.
\dontrun{
connect_to_limesurvey(
  api_url = 'https://localhost/index.php/admin/remotecontrol',
  limesurvey_username = 'admin',
  limesurvey_password = '1234admin',
  mysql_host = '127.0.0.1',
  mysql_port = 3306,
  mysql_dbname = 'limesurvey',
  mysql_username = 'lime',
  mysql_password = '1234lime'
)

wipe_survey_data(123456)
}

}
